% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PointLoc.R
\name{Point.Locator}
\alias{Point.Locator}
\title{Point.Locator
Point locator with ADTree, it works on 2D triangular mesh}
\usage{
Point.Locator(locations, num_loc, meshobj)
}
\arguments{
\item{locations,}{a vector containing the x-y coordinates of the points to search (x1,y1,x2,y2...)}

\item{num_loc,}{an integer containing the number of the point to be located}

\item{meshobj,}{the 2D mesh}
}
\value{
a vector containing the index of the triangle where the point are.
}
\description{
Point.Locator
Point locator with ADTree, it works on 2D triangular mesh
}
\examples{
library(rgl)
library(PointLoc)
nodes = rbind(c(0,0), c(10,0), c(10,10), c(0,10),cbind(runif(20, min = 0, max = 10), runif(20, min = 0, max = 10)))
mesh = create.MESH.2D(nodes, segments = rbind(c(1,2),c(2,3),c(3,4),c(4,1)))
mesh_2 = refine.MESH.2D(mesh,maximum_area = 1.)
mesh_2$order = 1
plot(mesh_2)
num_nodes = length(mesh_2$nodes)/2
loc <- c(2,3)
res = Point.Locator(loc, 1, mesh_2)
points(loc[1],loc[2], lwd = 3, col = 'green')
points(mesh_2$nodes[mesh_2$triangle[res,1],1], mesh_2$nodes[mesh_2$triangle[res,1],2], lwd = 3, col='blue')
points(mesh_2$nodes[mesh_2$triangle[res,2],1], mesh_2$nodes[mesh_2$triangle[res,2],2], lwd = 3, col='blue')
points(mesh_2$nodes[mesh_2$triangle[res,3],1], mesh_2$nodes[mesh_2$triangle[res,3],2], lwd = 3, col='blue')
lines(c(mesh_2$nodes[mesh_2$triangle[res,1],1], mesh_2$nodes[mesh_2$triangle[res,2],1]), c(mesh_2$nodes[mesh_2$triangle[res,1],2],mesh_2$nodes[mesh_2$triangle[res,2],2]), lwd = 3 ,col = "blue")
lines(c(mesh_2$nodes[mesh_2$triangle[res,3],1], mesh_2$nodes[mesh_2$triangle[res,2],1]), c(mesh_2$nodes[mesh_2$triangle[res,3],2],mesh_2$nodes[mesh_2$triangle[res,2],2]), lwd = 3 ,col = "blue")
lines(c(mesh_2$nodes[mesh_2$triangle[res,3],1], mesh_2$nodes[mesh_2$triangle[res,1],1]), c(mesh_2$nodes[mesh_2$triangle[res,3],2],mesh_2$nodes[mesh_2$triangle[res,1],2]), lwd = 3 ,col = "blue")
}

