% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Debug.R
\name{Test_ADTree}
\alias{Test_ADTree}
\title{Test for class ADTree
Test of some funtionalities on default ADTree (users can't modify them), I test all method I use or I modified: constructor, addtreenode, deltrenode, search, visualization of information}
\usage{
Test_ADTree(n)
}
\arguments{
\item{n}{is an int and is the input integer (n != 0)}
}
\value{
An int with value:
        \item{0}{if it's all ok}
        \item{1}{if there's some problem}
}
\description{
Test for class ADTree
Test of some funtionalities on default ADTree (users can't modify them), I test all method I use or I modified: constructor, addtreenode, deltrenode, search, visualization of information
}

