% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Debug.R
\name{Test_Domain}
\alias{Test_Domain}
\title{Test for class Domain
Test of some funtionalities on default domain (users can't modify them) with different shape, I test all method I use or I modified: constructor, operator <<, origin, scalingfactors, set and modify tolerance/mindiff}
\usage{
Test_Domain(n)
}
\arguments{
\item{n}{is an int and is the input integer (n != 0)}
}
\value{
An int with value:
        \item{0}{if it's all ok}
        \item{1}{if there's some problem}
}
\description{
Test for class Domain
Test of some funtionalities on default domain (users can't modify them) with different shape, I test all method I use or I modified: constructor, operator <<, origin, scalingfactors, set and modify tolerance/mindiff
}

