% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Debug.R
\name{Test_MeshHandler}
\alias{Test_MeshHandler}
\title{Test for class MeshHandler
Test of some funtionalities on default Mesh (users can't modify them), I test the constructor with the new flag and the search algorithm from ADTree}
\usage{
Test_MeshHandler(n)
}
\arguments{
\item{n}{is an int and is the input integer (n != 0)}
}
\value{
An int with value:
        \item{0}{if it's all ok}
        \item{1}{if there's some problem}
}
\description{
Test for class MeshHandler
Test of some funtionalities on default Mesh (users can't modify them), I test the constructor with the new flag and the search algorithm from ADTree
}

