% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Debug.R
\name{Test_TreeHeader}
\alias{Test_TreeHeader}
\title{Test for class Tree_Header
Test of some funtionalities on default Tree_Header (users can't modify them), I test all method I use or I modified: constructor, operator <<, get/change information on the TreeHeader}
\usage{
Test_TreeHeader(n)
}
\arguments{
\item{n}{is an int and is the input integer (n != 0)}
}
\value{
An int with value:
        \item{0}{if it's all ok}
        \item{1}{if there's some problem}
}
\description{
Test for class Tree_Header
Test of some funtionalities on default Tree_Header (users can't modify them), I test all method I use or I modified: constructor, operator <<, get/change information on the TreeHeader
}

