% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Debug.R
\name{Test_TreeNode}
\alias{Test_TreeNode}
\title{Test for class TreeNode
Test of some funtionalities on default treenode (users can't modify them), I test all method I use or I modified: constructor, print, set/get father/child/box/id/keys}
\usage{
Test_TreeNode(n)
}
\arguments{
\item{n}{is an int and is the input integer (n != 0)}
}
\value{
An int with value:
        \item{0}{if it's all ok}
        \item{1}{if there's some problem}
}
\description{
Test for class TreeNode
Test of some funtionalities on default treenode (users can't modify them), I test all method I use or I modified: constructor, print, set/get father/child/box/id/keys
}

