% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Debug.R
\name{Test_Triangle}
\alias{Test_Triangle}
\title{Test for class Triangle:
Test of some funtionalities on default triangle (users can't modify them), I test all the method I use or I modified: constructor, print, operator [], physical and search dimension, isPointInside}
\usage{
Test_Triangle(n)
}
\arguments{
\item{n}{is an int and is the input integer (n != 0)}
}
\value{
An int with value:
        \item{0}{if it's all ok}
        \item{1}{if there's some problem}
}
\description{
Test for class Triangle:
Test of some funtionalities on default triangle (users can't modify them), I test all the method I use or I modified: constructor, print, operator [], physical and search dimension, isPointInside
}

